+++ CHANGELOG.md
@@ -7,9 +7,9 @@
 - Bump Android SDK from v6.7.0 to v6.9.1 ([#406](https://github.com/getsentry/sentry-android-gradle-plugin/pull/406), [#408](https://github.com/getsentry/sentry-android-gradle-plugin/pull/408), [#411](https://github.com/getsentry/sentry-android-gradle-plugin/pull/411))
   - [changelog](https://github.com/getsentry/sentry-java/blob/main/CHANGELOG.md#691)
   - [diff](https://github.com/getsentry/sentry-java/compare/6.7.0...6.9.1)
-- Bump CLI from v2.8.1 to v2.9.0 ([#405](https://github.com/getsentry/sentry-android-gradle-plugin/pull/405))
-  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#290)
-  - [diff](https://github.com/getsentry/sentry-cli/compare/2.8.1...2.9.0)
+- Bump CLI from v2.8.1 to v2.10.0 ([#405](https://github.com/getsentry/sentry-android-gradle-plugin/pull/405), [#413](https://github.com/getsentry/sentry-android-gradle-plugin/pull/413))
+  - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#2100)
+  - [diff](https://github.com/getsentry/sentry-cli/compare/2.8.1...2.10.0)
 
 ## 3.3.0