+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - Do not register dependencies report task if it's disabled ([#422](https://github.com/getsentry/sentry-android-gradle-plugin/pull/422))
 - Ensure clean state before generating a new uuid by deleting the old `sentry-debug-meta.properties` file ([#420](https://github.com/getsentry/sentry-android-gradle-plugin/pull/420))
+- Support configuration cache in dependencies report task from Gradle `7.5` onwards ([#423](https://github.com/getsentry/sentry-android-gradle-plugin/pull/423))
 
 ### Dependencies